Put settings services in their own files
