+++ app/controllers/api/settings_controller.rb
@@ -12,45 +12,5 @@ module Api
       settings = SettingsSlicer.slice(whitelist, *path)
       render_resource :settings, settings
     end
-
-    private
-
-    class SettingsSlicer
-      def self.slice(settings, *path)
-        {}.tap { |h| h.store_path(path, settings.fetch_path(*path)) }
-      end
-    end
-
-    class SettingsFilterer
-      def self.filter_for(user)
-        new(user).fetch
-      end
-
-      attr_reader :user
-
-      def initialize(user)
-        @user = user
-      end
-
-      def fetch
-        whitelist_settings(settings_hash)
-      end
-
-      private
-
-      def whitelist_settings(settings)
-        return settings if user.super_admin_user?
-
-        result_hash = {}
-        ApiConfig.collections[:settings][:categories].each do |category_path|
-          result_hash.deep_merge!(SettingsSlicer.slice(settings_hash, *category_path.split("/")))
-        end
-        result_hash
-      end
-
-      def settings_hash
-        @settings_hash ||= Settings.to_hash.deep_stringify_keys
-      end
-    end
   end
 end
