+++ lib/services/api/settings_slicer.rb
@@ -1,0 +1,7 @@
+module Api
+  class SettingsSlicer
+    def self.slice(settings, *path)
+      {}.tap { |h| h.store_path(path, settings.fetch_path(*path)) }
+    end
+  end
+end
